Reduced walking of entire grid list
The collection is now a circular doubly linked list which means we can
always get the last from the first.

We do walk the entire list to count the number of them every time we
print the status every 30 seconds. We could just increment or decrement
a counter rather than doing this, but it's good enough for now.