Started a solver in C#
The idea being to use some of the new performance features.

The next-up idea being to use threading during the solving of a single grid.
I plan to do that by marking the spot in the queue where a worker thread is working.
If it does a push then it pushes behind it's marker.
If it does a pop it skips over ones that are marked.
All done when all threads exit (the queue should be empty).
This may not actually improve performance, we'll see.
Some care will be needed with the counters to avoid lock contention.

Might want to see if putting 'in' around the place improves performance.
Might be worth checking if the intrinsics actually help.
Might be worth seeing if avoiding '%' in red-routes via counting helps.