TS: Made a SudokuGrid and a linked list
Next I guess I can start the solver. I'm going to need to think what the contract for that will be.

Worth thinking about whether I also want a ValidSudokuGrid (which validates input e.g. x, y, value are within bounds). And if I want a ReadOnlyGrid so that users can see what's in the backtracking queue without being able to change it?

Does TypeScript or JavaScript have a way to make this in to an outer module? So that only the types needed by the user are exported?