Leds: Algorithm to render grid to LEDs
This is somewhat more of a reference implementation, but it should work in the Mega Processor with minor tweaks. For instance we'd want to copy the render target buffer to the memory area for the LEDs. We might also need to clear the buffer.

This doesn't include rendering `Unique`, `NotUniq` or `Imposs`, but those should be fairly simple.