VM: Decided calling convention
`r0` holds the stack frame pointer, which points to directly after the `pc` return address.

The first argument, if there is one, is passed in `r1`, otherwise `r1` is undefined.

The return result is passed in `r1`, otherwise `r1` is undefined.

For Jack method callers, this means the first arg, which is on the stack, is swapped in to `r1` and replaced with `r0` (or `r0` is pushed instead.

For Jack methods called this means that `r1` is pushed to the stack _before_ `sp` is moved to make space for locals and temp. This means that the first argument is loaded differently compared to other arguments.

I _think_ this means that `r0` always points to the first argument if there is one, which should be our grid pointer, which is the first argument to most methods. (I need to double check that)