ASM: Grid_isImpossible
I tweaked the memory layout in `.jack` to make this easier.

Maybe the grid cache should be at the start 'before' the grid pointer?
That would avoid the extra cycles of having to use `ld.w` rather than `ld.b`.
Not sure that's a change I can easily make given the `.jack`, but maybe!