ASM: `Input_acceptUserHints`
But at the moment it _never_ returns so there's no way to perform a solve or even to copy the values you set as hints in to the root grid.

The joystick has no obvious buttons that could be used to start the solve, so an on-screen button might be the best UX?

It might be best to get `SudokuSolver.asm` ready next.